Fix refresh breaking when story produces an error. (#25)
When a story throws an error, the live refresh fails until the user
reselects the story. This happens because the setup of the Changed
event occurs after parsing the story source. If parsing fails, then the
function returns before the Changed event is ever set up.

This change simply moves setup of the Changed event to occur before
parsing.